requirements.txt
aioconsole==0.8.2
fastapi==0.135.1
pydantic==2.12.5
pydantic_settings==2.13.1
python_bcrypt==0.3.2
python_jose==3.5.0
requests==2.32.5
SQLAlchemy==2.0.48
websockets==16.0
